requirements.txt
langchain
streamlit
python-dotenv
langchain_community
bs4
pypdf
langchain_text_splitters
langchain-ollama